Unifying tests for numcosmo and scipy.
