+++ tests/conftest.py
@@ -15,6 +15,7 @@ from firecrown.likelihood.gauss_family.statistic.statistic import TrivialStatist
 from firecrown.parameters import ParamsMap
 from firecrown.connector.mapping import MappingCosmoSIS, mapping_builder
 from firecrown.modeling_tools import ModelingTools
+from firecrown.models.cluster.abundance import ClusterAbundance
 
 
 def pytest_addoption(parser):
@@ -131,7 +132,7 @@ def fixture_tools_with_vanilla_cosmology():
 
 
 @pytest.fixture(name="cluster_sacc_data")
-def fixture_cluster_sacc_data():
+def fixture_cluster_sacc_data() -> sacc.Sacc:
     # pylint: disable=no-member
     cc = sacc.standard_types.cluster_counts
     # pylint: disable=no-member
@@ -155,3 +156,16 @@ def fixture_cluster_sacc_data():
     s.add_data_point(mlm, ("my_survey", "my_tracer2", "my_other_tracer1"), 1)
 
     return s
+
+
+@pytest.fixture(name="empty_cluster_abundance")
+def fixture_empty_cluster_abundance() -> ClusterAbundance:
+    cl_abundance = ClusterAbundance(
+        min_z=0,
+        max_z=2,
+        min_mass=13,
+        max_mass=17,
+        sky_area=100,
+        halo_mass_function=None,
+    )
+    return cl_abundance
